Refectoring and addressed comments
