+++ tools/setup_helpers/cuda.py
@@ -1,18 +1,13 @@
 import os
 import glob
 import re
-import platform
 import ctypes.util
 from subprocess import Popen, PIPE
 
-from .env import check_env_flag
+from .env import IS_WINDOWS, IS_LINUX, IS_DARWIN, check_env_flag
 
 LINUX_HOME = '/usr/local/cuda'
 WINDOWS_HOME = glob.glob('C:/Program Files/NVIDIA GPU Computing Toolkit/CUDA/v*.*')
-
-IS_WINDOWS = platform.system() == 'Windows'
-IS_LINUX = platform.system() == 'Linux'
-IS_DARWIN = platform.system() == 'Darwin'
 
 
 def find_nvcc():
